Move SDKv4 specific code to compat layer
